Improvement code readability
