+++ README.md
@@ -1,13 +1,35 @@
-# Simple Pinball
+# Pinball
 
 ###### _Authors: sgalella, mikirubio_
 
+#### Description
+
 We add our first own-written program. It was done during our first course in Programming and written in C.
 
-It consists in a simplification of a pinball game and basically has 4 options:
-###### 1.Visualizing the game table
-###### 2.Print some statistics about the table
-###### 3.Simulating 1 ball falling and watching the route it takes in the table
-###### 4.Simulating 10 balls fall and observing how many ball have gone through each position in the table
-	
-Remember to compile it using GNU (gcc) compiler
+It simulates the falling of a ball in a pinball. 4 different available options:
+1. Visualizing the game table
+2. Print some statistics about the table
+3. Simulating 1 ball falling and watching the route it takes in the table
+4. Simulating 10 balls fall and observing how many ball have gone through each position in the table
+
+#### Further work
+1. Add goal ending for the ball (game mode)
+2. Enhance visualization in an external window
+3. Generation of random pinballs
+
+
+#### Images
+
+Menu option
+
+![alt text](menu.jpg)
+
+Example of simulation of a ball fall. The numbers indicate the trajectory of the ball.
+
+![alt text](pinball.jpg)
+ 
+
+#### Notes
+Note: Remember to compile it using GNU (gcc) compiler.
+
+
