+++ statistics.h
@@ -1,0 +1,14 @@
+/*
+Project: Pinball
+Name: statistics.h
+Authors: mcasal & sgalella
+Description: compute statistics
+ */
+
+#include "amount_lines.h"
+#include "longest_line.h"
+
+void statistics (int table [50][50], int s[10]){
+	amount_lines(table,s);
+	longest_line(s);
+}
